services/auth-service/requirements.txt
alembic==1.14.*
asyncpg==0.30.*
psycopg2==2.9.*
sqlalchemy[asyncio]==2.0.*

pydantic==2.10.*
pydantic-settings==2.7.*

fastapi==0.115.*
uvicorn[standrd]==0.34.*
prometheus-fastapi-instrumentator==7.0.*

grpcio==1.67.*
grpcio-tools==1.67.*
werkzeug==3.0.*
python-jose[cryptography]==3.3.*
loguru==0.7.*